+++ src/effects/feats/blessedStrikes.js
@@ -2,7 +2,7 @@ import { baseFeatEffect } from "../specialFeats.js";
 
 export function blessedStrikesEffect(document) {
   if (document.system.actionType === null) return document;
-  let effect = baseFeatEffect(document, document.name);
+  let effect = baseFeatEffect(document, document.name, { transfer: true });
 
   effect.changes.push(
     {
@@ -27,7 +27,6 @@ export function blessedStrikesEffect(document) {
 
   document.system.damage.parts = [];
   document.system.actionType = null;
-  effect.transfer = true;
 
   document.effects.push(effect);
   return document;