+++ src/effects/feats/slayersPrey.js
@@ -4,7 +4,7 @@ import DDBMacros from "../DDBMacros.js";
 // this one is a bit different, the macro is triggered by midi-qol and applies effects to the actor
 // the Marked effect gets applied to the target
 export async function slayersPreyEffect(document) {
-  let effect = baseFeatEffect(document, `Marked by ${document.name}`);
+  let effect = baseFeatEffect(document, `Marked by ${document.name}`, { transfer: true });
   effect.changes.push(
     {
       key: "flags.dae.onUpdateSource",
@@ -13,20 +13,16 @@ export async function slayersPreyEffect(document) {
       priority: 20,
     },
   );
-  effect.transfer = false;
   effect.duration.seconds = 60;
   document.effects.push(effect);
 
-  let damageBonusEffect = baseFeatEffect(document, document.name);
+  let damageBonusEffect = baseFeatEffect(document, document.name, { transfer: true });
   damageBonusEffect.changes.push({
     key: "flags.dnd5e.DamageBonusMacro",
     value: "ItemMacro",
     mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM,
     priority: 20,
   });
-  damageBonusEffect.transfer = true;
-
-  setProperty(damageBonusEffect, "flags.dae.transfer", true);
   document.effects.push(damageBonusEffect);
 
   await DDBMacros.setItemMacroFlag(document, "feat", "slayersPrey.js");